+++ R/stat-alluvium.r
@@ -5,6 +5,7 @@
 #' intersections of the alluvia with the strata. It leverages the `group`
 #' aesthetic for plotting purposes (for now).
 #' @template stat-aesthetics
+#' @template computed-variables
 #' @template order-options
 #' @template defunct-stat-params
 #'
@@ -17,21 +18,20 @@
 #' @inheritParams stat_flow
 #' @param cement.alluvia Logical value indicating whether to aggregate `y`
 #'   values over equivalent alluvia before computing lode and flow positions.
-#'   Alternatively, a function (or its name) to combine the labels (if any) of
-#'   equivalent alluvia, similar to the `distill` parameter of
-#'   [to_alluvia_form()]). If set to `TRUE`, defaults to function
-#'   `[dplyr::first()]`.
 #' @param aggregate.y Deprecated alias for `cement.alluvia`.
 #' @param lode.guidance The function to prioritize the axis variables for
 #'   ordering the lodes within each stratum, or else a character string
 #'   identifying the function. Character options are "zigzag", "frontback",
 #'   "backfront", "forward", and "backward" (see [`lode-guidance-functions`]).
-#' @param lode.ordering A list (of length the number of axes) of integer vectors
-#'   (each of length the number of rows of `data`) or NULL entries (indicating
-#'   no imposed ordering), or else a numeric matrix of corresponding dimensions,
-#'   giving the preferred ordering of alluvia at each axis. This will be used to
-#'   order the lodes within each stratum by sorting the lodes first by stratum
-#'   and then by the provided vectors.
+#' @param lode.ordering **Deprecated in favor of the `order` aesthetic.** A list
+#'   (of length the number of axes) of integer vectors (each of length the
+#'   number of rows of `data`) or NULL entries (indicating no imposed ordering),
+#'   or else a numeric matrix of corresponding dimensions, giving the preferred
+#'   ordering of alluvia at each axis. This will be used to order the lodes
+#'   within each stratum by sorting the lodes first by stratum, then by the
+#'   provided vectors, and lastly by remaining factors (if the vectors contain
+#'   duplicate entries and therefore do not completely determine the lode
+#'   orderings).
 #' @example inst/examples/ex-stat-alluvium.r
 #' @export
 stat_alluvium <- function(mapping = NULL,
@@ -46,7 +46,7 @@ stat_alluvium <- function(mapping = NULL,
                           aggregate.y = NULL,
                           cement.alluvia = ggalluvial_opt("cement.alluvia"),
                           lode.guidance = ggalluvial_opt("lode.guidance"),
-                          lode.ordering = ggalluvial_opt("lode.ordering"),
+                          lode.ordering = NULL,
                           aes.bind = ggalluvial_opt("aes.bind"),
                           infer.label = FALSE,
                           min.y = NULL, max.y = NULL,
@@ -88,6 +88,9 @@ StatAlluvium <- ggproto(
   "StatAlluvium", Stat,
   
   required_aes = c("x"),
+  
+  # `<new-aes> = NULL` prevents "unknown aesthetics" warnings
+  default_aes = aes(weight = 1, stratum = NULL, alluvium = NULL),
   
   setup_params = function(data, params) {
     
@@ -163,18 +166,20 @@ StatAlluvium <- ggproto(
                            decreasing = ggalluvial_opt("decreasing"),
                            reverse = ggalluvial_opt("reverse"),
                            absolute = ggalluvial_opt("absolute"),
-                           discern = FALSE,
+                           discern = FALSE, distill = first,
                            negate.strata = NULL,
                            aggregate.y = NULL,
                            cement.alluvia = ggalluvial_opt("cement.alluvia"),
                            lode.guidance = ggalluvial_opt("lode.guidance"),
-                           lode.ordering = ggalluvial_opt("lode.ordering"),
+                           lode.ordering = NULL,
                            aes.bind = ggalluvial_opt("aes.bind"),
                            infer.label = FALSE,
                            min.y = NULL, max.y = NULL) {
     
     # introduce label
     if (infer.label) {
+      deprecate_parameter("infer.label",
+                          msg = "Use `aes(label = after_stat(lode))`.")
       if (is.null(data$label)) {
         data$label <- data$alluvium
       } else {
@@ -183,12 +188,32 @@ StatAlluvium <- ggproto(
       }
     }
     
-    # aesthetics (in prescribed order)
-    aesthetics <- intersect(.color_diff_aesthetics, names(data))
+    # ensure that `lode.ordering` is a matrix with column names
+    if (! is.null(lode.ordering)) {
+      deprecate_parameter("lode.ordering",
+                          msg = "Use the `order` aesthetic instead.")
+      if (is.null(data$order)) {
+        # bind a vector to itself to create a matrix
+        if (is.vector(lode.ordering)) {
+          lode.ordering <- matrix(lode.ordering,
+                                  nrow = length(lode.ordering),
+                                  ncol = length(unique(data$x)))
+        }
+        # flatten `lode.ordering` into an 'order' column
+        data$order <- as.vector(lode.ordering)
+      } else {
+        warning("Aesthetic `order` is specified, ",
+                "so parameter `lode.ordering` will be ignored.")
+      }
+    }
+    
+    # differentiation aesthetics (in prescribed order)
+    diff_aes <- intersect(c(.color_diff_aesthetics, .text_aesthetics),
+                          names(data))
     # match arguments for `aes.bind`
     if (! is.null(aes.bind)) {
       if (is.logical(aes.bind)) {
-        aes.bind.rep <- if (aes.bind) "flow" else "none"
+        aes.bind.rep <- if (aes.bind) "flows" else "none"
         warning("Logical values of `aes.bind` are deprecated; ",
                 "replacing ", aes.bind, " with '", aes.bind.rep, "'.")
         aes.bind <- aes.bind.rep
@@ -198,63 +223,100 @@ StatAlluvium <- ggproto(
     
     # sign variable (sorts positives before negatives)
     data$yneg <- data$y < 0
+    # lode variable (before co-opting 'alluvium')
+    data$lode <- data$alluvium
+    # specify distillation function from `distill`
+    distill <- distill_fun(distill)
+    
+    # initiate variables for `after_stat()`
+    weight <- data$weight
+    data$weight <- NULL
+    if (is.null(weight)) weight <- 1
+    data$n <- weight
+    data$count <- data$y * weight
+    # transform 'order' according to `absolute` and `reverse` params
+    if (! is.null(data$order)) data$order <- xtfrm(data$order) *
+      (-1) ^ (data$yneg * absolute + reverse)
     
     # cement (aggregate) `y` over otherwise equivalent alluvia
     if (! is.null(aggregate.y)) {
       deprecate_parameter("aggregate.y", "cement.alluvia")
       cement.alluvia <- aggregate.y
     }
-    if (is.logical(cement.alluvia)) {
-      label_fun <- first
-    } else {
-      if (is.character(cement.alluvia)) label_fun <- get(cement.alluvia)
-      stopifnot(is.function(cement.alluvia))
-      label_fun <- cement.alluvia
-      cement.alluvia <- TRUE
-    }
     if (cement.alluvia) {
-      data <- cement_data(data, "x", "alluvium", label_fun)
+      
+      # -+- need to stop depending on 'group' and 'PANEL' -+-
+      only_vars <- intersect(c(diff_aes, "group", "PANEL"), names(data))
+      bind_vars <- intersect(c("yneg", "stratum", only_vars), names(data))
+      sum_vars <- c("y", "n", "count")
+      
+      # interaction of all variables to aggregate over (without dropping NAs)
+      # -+- need to stop depending on 'group' -+-
+      data$binding <- as.numeric(interaction(lapply(
+        data[, bind_vars, drop = FALSE],
+        addNA, ifany = FALSE
+      ), drop = TRUE))
+      # convert to alluvia format with 'binding' entries
+      luv_dat <- alluviate(data, "x", "binding", "alluvium")
+      # sort by all axes (everything except 'alluvium')
+      luv_dat <- luv_dat[do.call(
+        order,
+        luv_dat[, setdiff(names(luv_dat), "alluvium"), drop = FALSE]
+      ), , drop = FALSE]
+      
+      # define map from original to aggregated 'alluvium' column
+      luv_orig <- luv_dat$alluvium
+      luv_agg <- cumsum(! duplicated(interaction(
+        luv_dat[, setdiff(names(luv_dat), "alluvium"), drop = FALSE],
+        drop = TRUE
+      )))
+      # transform 'alluvium' in `data` accordingly
+      data$alluvium <- luv_agg[match(data$alluvium, luv_orig)]
+      
+      # aggregate variables over 'x', 'yneg', and 'stratum':
+      # sum of computed variables and unique-or-bust values of aesthetics
+      by_vars <- c("x", "yneg", "stratum", "alluvium", "binding")
+      agg_lode <- stats::aggregate(data[, "lode", drop = FALSE],
+                                   data[, by_vars],
+                                   distill)
+      if (length(only_vars) > 0) {
+        agg_only <- stats::aggregate(data[, only_vars, drop = FALSE],
+                                     data[, by_vars],
+                                     only)
+      }
+      agg_dat <- stats::aggregate(data[, sum_vars],
+                                  data[, by_vars],
+                                  sum)
+      agg_dat <- merge(agg_dat, agg_lode)
+      if (length(only_vars) > 0) {
+        agg_dat <- merge(agg_dat, agg_only)
+      }
+      
+      # merge into `data`, ensuring that no `key`-`id` pairs are duplicated
+      data <- unique(merge(
+        agg_dat,
+        data[, setdiff(names(data), sum_vars)],
+        all.x = TRUE, all.y = FALSE
+      ))
+      data$binding <- NULL
+      
     }
     
     # define 'deposit' variable to rank strata vertically
     data <- deposit_data(data, decreasing, reverse, absolute)
     
-    # if `lode.ordering` not provided, calculate it (ignoring aesthetics)
-    if (is.null(lode.ordering)) {
-      # invoke surrounding axes in the order prescribed by `lode.guidance`
-      if (is.character(lode.guidance)) {
-        lode.guidance <- get(paste0("lode_", lode.guidance))
-      }
-      stopifnot(is.function(lode.guidance))
-      # summary data of alluvial deposits
-      alluv_dep <- alluviate(data, "x", "deposit", "alluvium")
-      # axis indices
-      alluv_x <- setdiff(names(alluv_dep), "alluvium")
-      # calculate `lode.ordering` from `lode.guidance`
-      lode.ordering <- matrix(NA_integer_,
-                              nrow = nrow(alluv_dep), ncol = length(alluv_x))
-      dimnames(lode.ordering) <- list(alluv_dep$alluvium, alluv_x)
-      for (xx in alluv_x) {
-        ord_x <- lode.guidance(length(alluv_x), match(xx, alluv_x))
-        # order by aesthetics in order
-        lode.ordering[, xx] <- interaction(alluv_dep[, alluv_x[rev(ord_x)]],
-                                           drop = TRUE)
-      }
-      # bind a vector to itself to create a matrix
-    } else if (is.vector(lode.ordering)) {
-      lode.ordering <- matrix(lode.ordering,
-                              nrow = length(lode.ordering),
-                              ncol = length(unique(data$x)))
+    # ensure that `lode.guidance` is a function
+    if (is.character(lode.guidance)) {
+      lode.guidance <- get(paste0("lode_", lode.guidance))
     }
-    # check that array has correct dimensions
-    stopifnot(dim(lode.ordering) ==
-                c(length(unique(data$alluvium)), length(unique(data$x))))
+    stopifnot(is.function(lode.guidance))
+    # invoke surrounding axes in the order prescribed by `lode.guidance`
+    lode_ord <- guide_lodes(data, lode.guidance)
     
-    # convert `lode.ordering` into a single sorting variable 'rem_deposit'
+    # convert `lode_ord` into a single sorting variable 'rem_deposit'
     # that orders index lodes by remaining / remote deposits
-    lode_ord <- as.data.frame(lode.ordering)
     names(lode_ord) <- sort(unique(data$x))
-    lode_ord$alluvium <- if (is.null(rownames(lode.ordering))) {
+    lode_ord$alluvium <- if (is.null(rownames(lode_ord))) {
       if (is.factor(data$alluvium)) {
         levels(data$alluvium)
       } else if (is.numeric(data$alluvium)) {
@@ -263,7 +325,7 @@ StatAlluvium <- ggproto(
         unique(data$alluvium)
       }
     } else {
-      rownames(lode.ordering)
+      rownames(lode_ord)
     }
     # match `lode_ord$x` back to `data$x`
     uniq_x <- sort(unique(data$x))
@@ -278,7 +340,7 @@ StatAlluvium <- ggproto(
     
     # identify fissures at aesthetics that vary within strata
     n_lodes <- nrow(unique(data[, c("x", "stratum")]))
-    fissure_aes <- aesthetics[which(sapply(aesthetics, function(x) {
+    fissure_aes <- diff_aes[which(sapply(diff_aes, function(x) {
       nrow(unique(data[, c("x", "stratum", x)]))
     }) > n_lodes)]
     data$fissure <- if (length(fissure_aes) == 0) {
@@ -289,13 +351,18 @@ StatAlluvium <- ggproto(
         (-1) ^ (data$yneg * absolute + reverse)
     }
     
-    # reverse alluvium order among negative observations
-    data$fan <- xtfrm(data$alluvium) * (-1) ^ (data$yneg * absolute + reverse)
+    # calculate variables for `after_stat()`
+    x_sums <- tapply(abs(data$count), data$x, sum, na.rm = TRUE)
+    data$prop <- data$count / x_sums[match(as.character(data$x), names(x_sums))]
+    
+    # reverse alluvium order
+    data$fan <- xtfrm(data$alluvium) * (-1) ^ reverse
     
     # sort data in preparation for `y` sums
     sort_fields <- c(
       "x",
       "deposit",
+      if (! is.null(data$order)) "order",
       if (aes.bind == "alluvia") "fissure",
       "rem_deposit",
       if (aes.bind == "flows") "fissure",
@@ -311,9 +378,10 @@ StatAlluvium <- ggproto(
       }
     }
     # calculate y bounds
-    data$deposit <- NULL
     data$rem_deposit <- NULL
+    data$order <- NULL
     data$fissure <- NULL
+    data$fan <- NULL
     data$ymin <- data$ycum - abs(data$y) / 2
     data$ymax <- data$ycum + abs(data$y) / 2
     data$y <- data$ycum
@@ -329,7 +397,7 @@ StatAlluvium <- ggproto(
     # (data is sorted by `x` and `alluvium`; group_by() does not reorder it)
     data <- dplyr::ungroup(dplyr::mutate(dplyr::group_by(data, alluvium),
                                          flow = axis - cumsum(cont)))
-    # add `group` to group contiguous alluvial subsets
+    # add 'group' to group contiguous alluvial subsets
     data <- transform(data, group = as.numeric(interaction(alluvium, flow)))
     # remove unused fields
     data$cont <- NULL
@@ -341,98 +409,40 @@ StatAlluvium <- ggproto(
     if (! is.null(max.y)) data <- subset(data, ymax - ymin <= max.y)
     
     # arrange data by aesthetics for consistent (reverse) z-ordering
-    data <- z_order_aes(data, aesthetics)
+    data <- z_order_aes(data, diff_aes)
     
     data
   }
 )
 
-# aggregate `y` and `label` over equivalent alluvia (omitting missing values)
-cement_data <- function(data, key, id, fun) {
+# apply lode guidance function to produce ordering matrix
+guide_lodes <- function(data, guidance_fun) {
   
-  agg_vars <- intersect(c("y", "label"), names(data))
+  # summary data of alluvial deposits
+  alluv_dep <- alluviate(data, "x", "deposit", "alluvium")
+  # axis indices
+  alluv_x <- setdiff(names(alluv_dep), "alluvium")
   
-  # interaction of all variables to aggregate over (without dropping NAs)
-  data$binding <- as.numeric(interaction(lapply(
-    data[, -match(c(key, id, agg_vars), names(data)), drop = FALSE],
-    addNA, ifany = FALSE
-  ), drop = TRUE))
-  # convert to alluvia format with 'binding' entries
-  alluv_data <- alluviate(data, key, "binding", id)
-  # sort by all axes (everything except `id`)
-  alluv_data <- alluv_data[do.call(
-    order,
-    alluv_data[, -match(id, names(alluv_data)), drop = FALSE]
-  ), , drop = FALSE]
+  # initialize ordering matrix
+  ord_mat <- matrix(NA_integer_,
+                    nrow = nrow(alluv_dep), ncol = length(alluv_x))
+  dimnames(ord_mat) <- list(alluv_dep$alluvium, alluv_x)
   
-  # define map from original to aggregated `id`s
-  alluv_orig <- alluv_data[[id]]
-  alluv_agg <- cumsum(! duplicated(interaction(
-    alluv_data[, -match(id, names(alluv_data)), drop = FALSE], drop = TRUE
-  )))
-  # transform `id` in `data` accordingly
-  data[[id]] <- alluv_agg[match(data[[id]], alluv_orig)]
-  
-  # aggregate `y` and `label` by all other variables
-  by_vars <- c(key, id, "binding")
-  data_agg <- dplyr::group_by(data[, c(by_vars, agg_vars)], .dots = by_vars)
-  data_agg <- if ("label" %in% agg_vars) {
-    dplyr::summarize(data_agg,
-                     y = sum(.data$y, na.rm = TRUE), label = fun(.data$label))
-  } else {
-    #dplyr::summarize(data_agg, y = sum(.data$y, na.rm = TRUE))
-    dplyr::summarize_at(data_agg, "y", sum, na.rm = TRUE)
+  # calculate orderings from `guidance_fun`
+  for (xx in alluv_x) {
+    ii <- match(xx, alluv_x)
+    ord_x <- guidance_fun(length(alluv_x), match(xx, alluv_x))
+    # order by prescribed ordering and by aesthetics in order
+    ord_mat[, xx] <- interaction(alluv_dep[, alluv_x[rev(ord_x)]],
+                                   drop = TRUE)
   }
-  data_agg <- dplyr::ungroup(data_agg)
-  # merge into `data`, ensuring that no `key`-`id` pairs are duplicated
-  data <- unique(merge(
-    data_agg,
-    data[, -match(agg_vars, names(data))],
-    all.x = TRUE, all.y = FALSE
-  ))
-  data$binding <- NULL
   
-  data
-}
-
-# aggregate weights over otherwise equivalent alluvia (omitting missing values)
-cement_data_alt <- function(data, key, id) {
+  # check that array has correct dimensions
+  stopifnot(dim(ord_mat) ==
+              c(length(unique(data$alluvium)), length(unique(data$x))))
   
-  # interaction of all variables to aggregate over (without dropping NAs)
-  data$.binding <- as.numeric(interaction(lapply(
-    data[, -match(c(key, id, "y"), names(data)), drop = FALSE],
-    addNA, ifany = FALSE
-  ), drop = TRUE))
-  
-  # convert to alluvia format with '.binding' entries
-  alluv_data <- alluviate(data, key, ".binding", id)
-  
-  # redefine 'alluvium' with only as many distinct values as distinct alluvia
-  alluv_agg <- stats::aggregate(
-    alluv_data[[id]],
-    by = alluv_data[, rev(setdiff(names(alluv_data), id))],
-    FUN = most
-  )
-  names(alluv_agg)[ncol(alluv_agg)] <- ".id"
-  alluv_data <- merge(alluv_data, alluv_agg, all.x = TRUE, all.y = FALSE)
-  data[[id]] <- alluv_data$.id[match(data[[id]], alluv_data[[id]])]
-  
-  # aggregate `y` by `key`, `id`, and .binding
-  data_agg <- stats::aggregate(
-    formula = y ~ .,
-    data = data[, c(key, id, ".binding", "y")],
-    FUN = sum
-  )
-  
-  # merge into `data`, ensuring that no `key`-`id` pairs are duplicated
-  data <- merge(
-    data_agg,
-    data[, -match("y", names(data))],
-    all.x = TRUE, all.y = FALSE
-  )
-  data$.binding <- NULL
-  
-  data
+  # return ordering matrix as a data aframe
+  as.data.frame(ord_mat)
 }
 
 # build alluvial dataset for reference during lode-ordering