+++ inst/examples/ex-geom-alluvium.r
@@ -8,9 +8,34 @@ ggplot(as.data.frame(Titanic),
 
 gg <- ggplot(alluvial::Refugees,
              aes(y = refugees, x = year, alluvium = country))
-# time series bump chart
+# time series bump chart (quintic flows)
 gg + geom_alluvium(aes(fill = country, colour = country),
-                   width = 1/4, alpha = 2/3, decreasing = FALSE)
+                   width = 1/4, alpha = 2/3, decreasing = FALSE,
+                   curve = "quintic")
 # time series line plot of refugees data, sorted by country
 gg + geom_alluvium(aes(fill = country, colour = country),
                    decreasing = NA, width = 0, knot.pos = 0)
+
+# irregular spacing between axes of a continuous variable
+refugees_sub <- subset(alluvial::Refugees, year %in% c(2003, 2005, 2010, 2013))
+gg <- ggplot(data = refugees_sub,
+             aes(x = year, y = refugees, alluvium = country)) +
+  theme_bw() +
+  scale_fill_brewer(type = "qual", palette = "Set3")
+# proportional knot positioning (default)
+gg +
+  geom_alluvium(aes(fill = country),
+                alpha = .75, decreasing = FALSE, width = 1/2) +
+  geom_stratum(aes(stratum = country), decreasing = FALSE, width = 1/2)
+# constant knot positioning
+gg +
+  geom_alluvium(aes(fill = country),
+                alpha = .75, decreasing = FALSE, width = 1/2,
+                knot.pos = 1, knot.prop = FALSE) +
+  geom_stratum(aes(stratum = country), decreasing = FALSE, width = 1/2)
+# coarsely-segmented curves
+gg +
+  geom_alluvium(aes(fill = country),
+                alpha = .75, decreasing = FALSE, width = 1/2,
+                curve = "arctan", segments = 6) +
+  geom_stratum(aes(stratum = country), decreasing = FALSE, width = 1/2)